requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-multipart>=0.0.6
google-generativeai>=0.3.0
requests>=2.31.0
python-magic>=0.4.27
aiofiles>=23.2.1
python-magic-bin>=0.4.14; platform_system=="Windows" 